New Features (Patch 1.18)
* Implemented the job listings
You can now add job listings to the website.

*Algorithm is 100% (To be Q.A'd)
Algorithm now responds to applicant skills to job categories and roles.
It can determine if it matches your desired content.

* Added Company Profile Edit (ongoing feature still has two missing features)
You can edit the company profile and add profile pictures that can be displayed on the banner. (Planning to display it on the job webpage as a cool little aesthetic feature.